goalconvo-backend/requirements.txt
# Core dependencies
transformers>=4.30.0
torch>=2.0.0
bert-score>=0.3.13
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.4.0
scikit-learn>=1.3.0

# Optional for advanced features
datasets>=2.12.0  # For MultiWOZ dataset
huggingface-hub>=0.16.0  # For model downloads

# Backend server dependencies
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
